Revert "Merge remote-tracking branch 'origin/beta' into beta-dev"
This reverts commit 5753777caeff015b8597a0c71ef3ce0fdd22a71d, reversing
changes made to 615e22ea8609e25a2a48f96303018e2b56fa0d96.
