+++ .gitignore
@@ -8,3 +8,4 @@ __pycache__
 data/user_templates/*.json
 resources/*.lib
 resources/venv
+ressources/
