+++ resources/mymodbus/mymodbusd.py
@@ -1,30 +1,10 @@
-import json
 import asyncio
 
-from jeedomdaemon.base_daemon import BaseDaemon
-from jeedomdaemon.base_config import BaseConfig
-from jeedomdaemon.utils import Utils
+from jeedomdaemon.base_daemon import BaseDaemon # type: ignore
 
 from mymodbusclient import MyModbusClient
+from mymodbusconfig import MyModbusConfig
 
-# --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-
-class MyModbusConfig(BaseConfig):
-  def __init__(self):
-    super().__init__()
-
-    self.add_argument("--json", help="MyModbus configuration json string", type=str, default='{}')
-
-  @property
-  def json(self):
-    """Returns the decoded json."""
-    return json.loads(self._args.json)
-  
-  @json.setter
-  def json(self, config):
-    self._args.json = json.dumps(config)
-
-# --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 
 class MyModbusd(BaseDaemon):
   def __init__(self) -> None:
